+++ checks/prgenv/src/hello/hello.c
@@ -1,0 +1,1 @@
+hello.cpp
