--- checks/prgenv/src/openmp/hello_world_openmp.c
@@ -1,16 +1,0 @@
-#include <omp.h>
-#include <stdio.h>
-#include <stdlib.h>
-
-int main (int argc, char *argv[]) {
-  int tid, nthreads;
-
-  #pragma omp parallel private(tid, nthreads)
-  {
-    tid = omp_get_thread_num();
-    nthreads = omp_get_num_threads();
-	printf("Hello, World from thread %d out of %d from process %d out of %d\n",
-		tid, nthreads, 0, 1);
-  }
-  return EXIT_SUCCESS;
-}
