--- checks/prgenv/src/serial/hello_world_serial.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-
-int main()
-{
-    printf("Hello, World from thread %d out of %d from process %d out of %d\n", 0, 1, 0, 1);
-
-    return 0;
-}
