--- checks/prgenv/src/serial/hello_world_serial.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main()
-{
-    std::cout << "Hello, World from thread 0 out of 1 from process 0 out of 1\n";
-}
