--- checks/prgenv/src/serial/hello_world_serial.f90
@@ -1,5 +1,0 @@
-program hellow
-implicit none
-write (*,'(A24,1X,I3,1X,A6,1X,I3,1X,A12,1X,I3,1X,A6,I3)') 'Hello, World from thread', 0, &
-	'out of', 1, 'from process', 0, 'out of', 1
-end
